% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_relative.R
\name{day_relative}
\alias{day_relative}
\title{Get Daily AT Internet Data - Relative Period -
This function allows you to get daily data from 'AT Internet' using relative period. Relative periods allow you to use periods that are automatically updated based on the last minute of available data. It is no longer necessary to change the date, week, etc : you simply need to specify which month, week, etc. you want to refer to for this last minute of available data.
METRICS are visits, visitors, conversions visits, conversions rate, visits with sales, time spent per pages, bounce rate}
\usage{
day_relative(ATuser, ATpassword, space, period_duration,
  dim_source = NULL, dim_device = NULL, dim_level2 = NULL,
  segment = NULL, pagination = 1)
}
\arguments{
\item{ATuser}{Your 'AT Internet' user login}

\item{ATpassword}{Your 'AT Internet' password login}

\item{space}{Analysis space}

\item{period_duration}{Analysis period}

\item{dim_source}{Optional - DIMENSION Trafic Source : Direct+traffic, Search+engines, "Referrer+sites", "Social+Campaigns"}

\item{dim_device}{Optional - DIMENSION Device Type}

\item{dim_level2}{Optional - DIMENSION Sub-set of your site}

\item{segment}{Optional - SEGMENT applied to the result set}

\item{pagination}{By defaut the pagination is set to 1. This allows you to get 10,000 rows of data. If you need more, you must make multiple calls to the API using the pagination feature.}
}
\value{
day_relative extract the data you need from 'AT Internet' database, according to parameters you entered and then provides a dataset.
}
\description{
Get Daily AT Internet Data - Relative Period -
This function allows you to get daily data from 'AT Internet' using relative period. Relative periods allow you to use periods that are automatically updated based on the last minute of available data. It is no longer necessary to change the date, week, etc : you simply need to specify which month, week, etc. you want to refer to for this last minute of available data.
METRICS are visits, visitors, conversions visits, conversions rate, visits with sales, time spent per pages, bounce rate
}
\examples{
\dontrun{
ATuser <- "user_email_adress"
ATpassword <- "my_password"
space <- 113736
period_duration <- 90
dim_source <- "Search+engines"
dim_device <- "Desktop"
dim_level2 <- "MyLevel2"
segment <- 204529097
pagination <- 1

dataset <- day_relative(ATuser, ATpassword, space, period_duration, dim_source, dim_device, dim_level2, segment, pagination)
}

}
