% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_set.R
\name{day_set}
\alias{day_set}
\title{Get Daily AT Internet Data - Set Period}
\usage{
day_set(ATuser, ATpassword, space, start_date, end_date, dim_source,
  dim_device, dim_level2, segment, pagination)
}
\arguments{
\item{ATuser}{Your 'AT Internet' user login}

\item{ATpassword}{Your 'AT Internet' password login}

\item{space}{Analysis space}

\item{start_date}{Start date}

\item{end_date}{End date}

\item{dim_source}{Optional - DIMENSION Trafic Source : Direct+traffic, Search+engines, "Referrer+sites", "Social+Campaigns"}

\item{dim_device}{Optional - DIMENSION Device Type}

\item{dim_level2}{Optional - DIMENSION Sub-set of your site}

\item{segment}{Optional - SEGMENT applied to the result set}

\item{pagination}{By defaut the pagination is set to 1. This allows you to get 10,000 rows of data. If you need more, you must make multiple calls to the API using the pagination feature.}
}
\value{
day_set extract the data you need from 'AT Internet' database, according to parameters you entered and then provides a dataset.
}
\description{
This function allows you to get daily data from 'AT Internet' using set perdiod, with a start date and an end date.
METRICS are visits, visitors, conversions visits, conversions rate, visits with sales, time spent per pages, bounce rate
}
\examples{
\dontrun{
ATuser <- "user_email_adress"
ATpassword <- "my_password"
space <- 113736
start_date <- "2018-11-01"
end_date <- "2018-11-10"
dim_source <- "Search+engines"
dim_device <- "Desktop"
dim_level2 <- "MyLevel2"
segment <- 204529097
pagination <- 1

dataset <- day_set(ATuser, ATpassword, space, start_date, end_date, dim_source, dim_device, dim_level2, segment, pagination)
}

}
